seperating db and server
