+++ azure/index.js
@@ -3,7 +3,6 @@ var port = process.env.PORT || 8080;
 var express = require("express");
 var app = express();
 app.use(require("body-parser").json());
-var connectMongoDB = require("./database.js").connectMongoDB;
 
 var ownerCommands = [];
 var visitorRequests = [];
@@ -59,40 +58,7 @@ app.post("/azure/visitor", function(req, res) {
   res.send(response);
 });
 
-function testInsert(client) {
-  var db = client.db("keyless");
-  console.log(db);
-  var collection = db.collection("devices");
-  collection
-    .findOne({ name: "test" })
-    .then(function(res, rej) {
-      if (res) {
-        console.log(res);
-      } else {
-        collection.insertOne({ name: "test", password: "test" }, function(err, res) {
-          if (err) {
-            console.log(err);
-            return;
-          }
-          console.log(res);
-        });
-      }
-      client.close();
-    })
-    .catch(function(err) {
-      console.log(err);
-      client.close();
-    });
-}
-
-connectMongoDB()
-  .then(function(client) {
-    app.listen(port, function() {
-      console.log("Example app listening on port " + port + "!");
-      testInsert(client);
-    });
-  })
-  .catch(function(err) {
-    console.log(err);
-    console.log("Couldn´t start server, caused by above problem");
-  });
+app.listen(port, function() {
+  console.log("Example app listening on port " + port + "!");
+  testInsert(client);
+});
